app/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Eugene Cho <eugenecho@cohere.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
cohere = "^5.13.11"
openai = "^1.61.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
